# [Merged by Bors] - Filter material handles on extraction