@alice-i-cecile 2022-03-25
> this wouldn't really be improved by https://github.com/bevyengine/bevy/pull/4244: the computed_visibility.is_visible step is a filter which would lose the exact size iterator property of a known size

Ah yep, you're right there. Sorry for misleading!