@superdump 2022-05-03
I've retested after the `extract_visible()` constructor suggestion with `many_cubes -- sphere` for the extraction system and its corresponding system commands spans.
main: 0.519ms + 3.52ms
this: 0.386ms + 0.221ms

Not sure why extraction of StandardMaterial handles is now about 0.1ms slower than I measured before, but it's still faster and way faster system commands as most of the entities are not visible.